TIFFFormat: remove unnecessary "init" methods
There was still a method called "initMetadataStore" which was leftover
from Bio-Formats, as there is no "MetadataStore" class in SCIFIO.
Further, it performs operations that are not worth distinguishing from
"initStandardMetadata" now that calibration values are part of the core
SCIFIO Metadata model.

Thus, the "initStandardMetadata" and "initMetadataStore" methods were
removed and their code placed into "initMetadata" method of the
BaseTIFFParser class.