FakeFormat: fix color table data structures
Color tables in FakeFormat should be generated to match the
getColorTable api, which indexes on imageIndex and planeIndex.
FakeFormat had been relying on the old logic of one lut per Channel
axis as was the case in Bio-Formats, but this logic was outdated and
would cause errors when the length of that axis and the number of planes
varied.

So the ColorTable data structure has been refactored to a 2D array
indexed on imageIndex and planeIndex. All uses of this structure have
been simimlarly converted, to remove the use of Channel axis
length/position.

This updates the FakeFormat.Metadata api to getLuts and setLuts. The old
methods have been preserved and deprecated, now operating only on image
index 0.

Added unit tests to verify the number of tables being created in
different indexed color situations.

Closes #116