TiffParser: fix IFD entry truncation
The TiffParser#getIFD method had logic for truncating IFD entries that
would overflow the input. However, although a truncated length was
calculated, it was never actually used.

Updated to replace the overflowing IFD length with the truncated
length. Now, instead of throwing an exception when encountering an IFD
entry that indicates reading past the EOF, it will be read as much as
possible and handled gracefully.

Closes #123