AWTImageTools: remove protected field
There was no reason for this to be protected - there are no
AWTImageTools subclasses. If it was intended to be package scoped, that
is terrible. If a use case arrises, this can just be made public,
because it is static final (or an accessor which creates a deep copy, or
something.. anything other than package protected).