AVIFormat: fix initialization
Because of the effectiveAxis logic, adding a length 1 axis without
adjusting the planar axis count is effectively accomplishing nothing,
thus attempting to set the axis type when that axis is not really
considered "present" results in AIOOB exceptions.

Solution is to set the planar axis count, and then set the axis type.