SCIFIOCellImg closes the reader upon disposal
SCIFIOCellImgs are odd cases in that they need to remain open for the
full life of the ImgPlus they are attached to, to allow dynamic loading
of plane regions.

To facilitate this, a SCIFIOCellImgCleaner was created, and
SCIFIOCellImgs now register themselves for cleanup on creation.

SCIFIOCellImg and SCIFIOCellImgPlus also both now implement disposable,
allowing manual disposal.

Closes #128