Convert command: limit plane processing by writer
Convert operates as a Reader command, so the number of planes processed
is controlled by the number of planes on the Reader. However, Convert is
a special case since it also has a Writer. There are cases where planes
will be truncated (e.g. writing to EPSFormat from a source with time
points). So on the Convert side of the processPlane workflow, nothing
should be done if beyond the Writer's plane limit.