ICSFormat: do not add lifetime axis of length 1
SCIFIO's logic recently changed such that it now aggressively prunes
axes of length 1. But the ICS parser was adding a lifetime axis of
length 1, then later updating the length once more information was
known. Unfortunately, the axis pruning interfered with this approach.

It could be argued that SCIFIO should be less militant about axis
pruning; i.e., perhaps it should still allow referencing axes by index
even if they are internally of length 1 and hence not included when
reporting the dimensionalty.

That said, the fix in this case is simple: do not add an axis of length
1 even temporarily, but instead add the axis with the proper known
length in the first place.