ICSFormat: simplify Gray Institute lifetime hack
The previous hack was migrated from Bio-Formats, and was laden with a
few assumptions that no longer hold. It also caused problems because in
the case of "t x y" the dimensionality ended up being
[X, Y, Z, Lifetime] with Z=1 (i.e., an embedded 1-length dimension).

The new hack is shorter, and works simply by swapping axis types around
as appropriate. It also purges the no-longer-needed planar axis count
computation, since the default planar axis guessing logic will always be
correct in the case of ICS.