MinimalTIFFFormat: correct IFD index
During the openPlane call in MinimalTIFFFormat, the (planeNo) index of
the IFD list was being accessed. However, this is not guaranteed to
exist.

The correct index should be the (imageNo) index of the IFD list.

This will allow the opening of some TIFF datasets that were previously
failing with SCIFIO.