AVIFormat: account for non-planar Channels
The AVIFormat folds in existing channel axes to planar axes, to create
RGB data. However, in the case that Axes.CHANNEL was not a planar axis
in the source dataset, the channels were being counted twice.

This has been corrected by factoring out the Axes.CHANNEL length from
the plane count, and including it in plane length, when applicable.