Fix cached plane logic
ChannelFiller and PlaneSeparator had erroneous logic in determining if
they had a cached plane or not (if the requested region contained the
cached region, the cached plane was returned - instead of the intended
reverse).

However, as there was no logic to properly extract desired sub-regions
from cached planes, cached planes will only be returned for now when
they match exactly.

This functionality will be reviewed in the future
(see https://github.com/scifio/scifio/issues/155)