ImgOpener: allow opening of all images
ImgOpener now returns a list of ImgPlus objects, and can be instructed
to open specific ranges or all known images in a dataset through
configuration with the SCIFIOConfig class.

This required updates to the SCIFIOCellImg so that cells are now
associated with a particular index, to ensure the proper data is opened
when requested.

Added unit tests using .fakes to ensure the new functionality is working
as intended. This is a breaking API change.