AbstractImageMetadata: fix setAxis
When replacing axes (e.g. the new axis type was not found, so its index
was -1), was trying to use the axis index (of -1) in the axis array..
which of course was invalid.

Corrected the logic so that if the axis already exists in the axis list,
its index is used, otherwise it replaces the axis at the given index.