Breaking news: Get rid of TestNG in favor of JUnit
This commit intentionally breaks all of the unit tests, by removing the
TestNG dependency. The changes required to make the unit tests work with
JUnit are rather extensive, and it is more important to group them into
small, reviewable commits (also for documentation purposes to help port
other branches to JUnit) than to keep the tests working.

Therefore, this commit does nothing else than to remove the TestNG
dependency and convert the obvious imports.

Most notably, the TestNG grouping (and running tests with a couple of
parameters) will be solved separately in a later commit (because there
is no easy JUnit equivalent for it).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>