Fall back to old FLIM hack if no micro-time data
The updated micro-time ICS files from Gray still have the history labels
tags that triggered the old ICS hack (and they are required for reading
on the Gray side). But the old ICS hack is incompatible with the new
micro-time data (because micro-time data does not require a
dimensionality hack, it just's a signal to use the lifetime axis label)
so we need to avoid using it when micro-time labels are detected.