LocationService: don't warn on VirtualHandle
There are times when a VirtualHandle is exactly what we expect to create
(e.g. for FakeFormat or OMEROFormat).

So instead of logging warnings, just continue as normal. As
VirtualHandle now throws exceptions on unintended functionalities, we
will get appropriate failures if someone does try reading or writing to
the handle.