ImgSaver: eliminate redundant writer population
If writing via SCIFIO we were always populating metadata twice. Instead,
we should propagate null Writers and if null, perform initialization
once.

Similarly, if the Writer doesn't have attached Metadata, it is generated
and populated using the given ImgPlus.

If given a fully inintialized writer, we just leave it alone.