Set the ImgPlus source to the intended destination
This partially reverts commit 59322cf2fc1e9bba4cf581f919920fc198472cf7.

It is a huge hack -- as well as destructive -- to overwrite the
ImgPlus's source just so that the writer can extract that information
later. It would be better to find a nicer way to pass the destination to
the writer via the ImgPlus. We could use a property in the properties
table, but that would have the same problem of being vulnerable to race
conditions when attempting to write the image to multiple destinations
at the same time. There must be a better way...