ByteArrayHandle: don't null buffer on close
It is common that after writing to a ByteArrayHandle we will want to
obtain the array that was populated. However, if the handle is closed
its buffer was being nulled out. This is unnecessary, as when the handle
is discarded its buffer will be as well.

Similar to the functionality of ByteArrayOutputStream.close