MetadataWrapper: delegate getFormat calls
For informative purposes we should really return the actual format
associated with a wrapped Metadata. We were returning null.

Although it may be confusing that
MetadataWrapper.getFormat.createMetadata will not create the same type
of metadata as you started with, I think it is worth it to avoid having
a null Format - which precludes you from creating readers, writers,
etc...