SCIFIOArrayLoader: add loadTable method
When working with Imgs and ImgPluses, we need to preserve ColorTables
opened by SCIFIO (which are attached automatically to planes). So at the
plane-opening layer, we need to be able to cache and return associated
ColorTables.

The new API provided is intended to return cached ColorTables that were
saved while opening planes, or load them if the plane wasn't previously
opened. This should minimize unnecessary plane opening of
SCIFIOCellImgs.