Don't check indexed state for ImgPlus
We were checking if the base image was indexed to determine the
composite channel count (the base had to be checked because we're always
using channel fillers and plane separators, so the top-level metadata
shouldn't have any planar channels).

First of all, we should have been setting the # of composite channels to
the LUT length in cases of indexed color, and second we should have been
also checking if the image was false color (in which case there are NOT
LUT length # of additional real channels).

But for now since we are using the ChannelFiller, let's just remove the
index basis for setting composite channel count.