PlaneSeparator: account for negative indices
When determining indices of separated axes, we were just checking if the
index was less than the number of planar axes. However, if the index was
-1 it was of course less than the number of planar axes, yet we wanted
to trigger index population as though the plane had been separated
(because a negative index suggests that a length 1 axis was separated
and there were no significant trailing axes, resulting in its discard).