TIFFFormat: parse IJ1 color table
ImageJ 1.x can write 8-bit lookup tables in the TIFF ImageJ 1.x comment.
These are not properly read when retrieved from the IFD as a String.
They need to be parsed by reading the metadata tags and determining the
actual offsets and lengths.

We now parse and use the IJ 1.x color table if needed.

Closes #202

NB: See also #203 for future plans