FormatTools: fail loudly when axis is nonlinear
Previously, the calibration would fail silently. But in this scenario we
really want to know if something went wrong, including if the axis in
question is null.

Time will tell whether this change is too aggressive. If that proves to
be the case, we can perhaps return a boolean instead indicating success.
But doing nothing silently is bad news.