TIFFFormat: apply the spacing to the correct axis
Before this change, the spacing would be parsed but then lost, because
at the time of parsing, there are three axes: X, Y and TIME. There is no
Z, so a new Z is created. But later the axes get reshuffled and that Z
axis gets lost again, including its unit and spacing, in favor of the
time axis (which was never calibrated properly).

This change defers the assignment of the calibration until the axes have
settled down, resulting in a correctly calibrated Z axis.

This is not the greatest patch in the world, but it gets the job done
for now, until the ImageJ comment parsing logic gets rewritten as
documented in issue #203.

This change address the second half of Fiji bug 802.