MetadataTest: switch to JUnit
Running TestNG tests in Eclipse is a real pain. Let's use a framework
that actually works out of the box.

Note that this conversion exposed at least one lurking bug in the unit
tests: a test method with "expectedExceptions" is throwing its
exception _before_ the comment in the source says it should. But it was
not noticed due to how that annotation behaves with TestNG...