AbstractImageMetadata: fix bug in setAxisLength
The problem, identified by MetadataTest#testAdjustTrailingAxis, was that
the getAxisIndex method now discards trailing axes of length 1. So the
setAxisLength method was erroneously determining that it needed to add
another axis of the same type that already existed in the raw axes list.

The fix is to check the _raw_ axes list for the axis of that type,
rather than the "effective" axes list as was being done.