SCIFIOConfig: writer compression default is null
An empty string is almost certainly not a valid compression type for any
format. It also can cause problems if downstream code is checking for
null values (specifically in the OME-XML component).

Thus the config will now default compression to null.

Updated the Writer javadoc to make this expectation explicit.

Updated TIFFFormat to add a defensive null check (every other consumer
of the getCompression method checked for nulls already).