PlaneSeparator: fix y index
The parent and PlaneSeparator may have different Y-axis indices, if an
axis that precedes Y is separated out. Thus the parentYIndex should be
used for accessing the planar offsets being sent to the parent, but the
PlaneSeparatorMetadata's Y-index should be used for indexing into the
offset array provided to it.

Reported by Christian Dietz.