DatasetIOPlugin: do not rely on deprecated methods
The DatasetService is the wrong thing to lean on for opening and saving
data here, especially since the DefaultDatasetService implementation
currently invokes SCIFIO by reflection anyway.

This change simply uses the SCIFIO-based logic as is to implement the
methods. At some point in the future, we may want a dedicated service
for this (e.g., io.scif.io.DatasetIOService), but for now, it is enough
to have the implementation in the DatasetIOPlugin directly. There will
be substantial structural code changes in the near future anyway, with
the merger of the scijava-common robust-io work.