Always obtain codecs from the CodecService
Codecs are singleton plugins. The idea is to only ever create one
instance, and then reuse that instance, since it is stateless.

Actually, singleton plugins are not completely stateless: they sometimes
need to know their context. Creating a new codec via its no-args
constructor will not populate the context by default, which is why in
some places the construction was followed up with a setContext call.

But it's better to simply ask the CodecService for the codec in
question; that way, the CodecService behavior could potentially be
cleanly overridden.

This addresses a problem with opening DICOM files (and probably certain
other file formats, too) reported on the ImageJ mailing list.