Remove EncodingTest
It was not clear what the EncodingTest tried to test, and neither
whether it does so correctly.

For example, testing whether the letter 'ß' is either encoded correctly as
UTF-8 when given the bytes corresponding to MacRoman encoding, but
silently succeeding if either MacRoman or UTF-8 are unknown encodings, is
most likely incorrect: new String(String, String)'s behavior is
unspecified when the given bytes are invalid in the given encoding.

So let's remove the test for the time being. It can be readded later
should the need arise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>