Convert TestNG grouping into a JUnit equivalent
TestNG has a feature where you can run a bunch of tests with a list of
parameters. To replicate the same with JUnit, we make use of the
Parameterized runner (which is also handled nicely by Eclipse) that
allows specifying a list of constructor arguments that will be used to
instantiate the unit test several times.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>