assertEquals(double, double) requires delta now
We need to adjust for the fact that the JUnit version we use no longer
lets assertEquals(double, double) pass: it complains that it requires a
delta to account for the fact that double precision values are
imprecise.

We introduce a new helper function to test two double (or single)
precision values whether they are "close enough", i.e. whether they
differ at most in two bits of their mantissa (which is actually pretty
precise, still).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>