Initialize Formats on separate thread
Instead of lazily generating the Format mappings, we queue their
construction on a separate thread. This could hurt worst-case-scenario
performance (when there isn't another thread to run on) but in general
should mitigate performance hits on initial calls into the
FormatService.

Note that this is a temporary workaround that can be made much better
with the completion of
https://github.com/scijava/scijava-common/issues/145