Set Dataset.isRGBMerged appropriately
The DatasetService.createDataset makes some assumptions which we
would like to circumvent regarding whether or not the Dataset should be
RGBMerged.

For now we will just check the -base- Metadata (ignoring any
PlaneSeparator or ChannelFiller activity) and look for Axes.CHANNEL in
the planar axis list. If found, the Dataset is RGBMerged.