WIP: POM: bump to scijava-common 3.0.0
This updates the scijava-common library dependency to 3.0.0,
which has some breaking changes from the 2.x versions. In particular,
it overhauls the I/O framework, changing the API contract of IOPlugins.
So this commit makes the necessary code changes to DatasetIOPlugin to
account for those. Later, SCIFIO will be upgraded more thoroughly to
leverage the improved SciJava Common I/O framework, to truly
generalize the I/O paradigm beyond only files on disk.