Correct interpretation of AT DICOM attributes
Before this change DICOM attributes with value representation AT
(attribute tag, which means pointers to other attributes using their
tags) are interpreted as strings (and then trimmed) causing meaningless
results.

With this change the value is interpreted as a list of couples [tag
group, tag element] and a user-friendly string is generated; the stirng
is a sequence of 8-characters-long substrings "<tag group><tag element>"
where <tag group> and <tag elements> are 4 characters long each.

Example
Image with value 0x18 0x00 0x63 0x10 (little endian) for the attribute
0028,0009 Frame Incremen Pointer without this change produces the string
"c" and with this change the string "00181063".