MicromanagerFormat: require companion files
The Checker now returns false if just given a RAIS, as it is not
possible to know if there's a companion metadata file right now.

Similarly, the Parser will now fail if there is no companion metadata
file.

This should reduce confusion in format matching.

See the related forum post raising this issue:
http://forum.imagej.net/t/reading-tiffs-directly-from-inputstream/462/5

And the robust fix for fileless reading:
https://github.com/scijava/scijava-common/issues/215