DICOMFormat: fix NPE when group=true
Closes #291.

The error being fixed:

    java.lang.NullPointerException
        at io.scif.formats.DICOMFormat$Parser.addFileToList(DICOMFormat.java:1642)

The code was essentially doing the following:

    private void buildFileList() {
      if (metadata.getFileList() == null) {
        fileList = new Hashtable();
        buildUpTheFileList();
        metadata.setFileList(fileList);
      }
    }
    private void buildUpTheFileList() {
      metadata.getFileList().add(matchingFile());
    }

In other words: the helper methods of buildFileList assumed that the
file list was already injected into the metadata table, when it wasn't.
And of course this threw an NPE.

The fix is to pass the local file list data structure to applicable
helper methods.

(The other fix would have been to set the not-yet-fully-baked
file list to the metadata right away, but that felt too dirty.)