DICOMFormat: split dictionary to its own class
This encapsulates the huge DICOM dictionary, so that it can be easily
queried by the DICOM format itself. I think this makes the DICOM
format much easier to read and understand, since there are not 4000
lines of dictionary-building code clogging up the source anymore.

This change also encapsulates the implementation details of the
dictionary; for example, if we move away from java.util.Hashtable
in the future, it will not be a breaking change, since the
DICOMDictionary public API is minimal.