Fix bogus "expected positive value" error messages
This is a really old, and really dumb, bug: when [XY]Resolution is
absent in the TIFF (which is the case extremely often), the
IFD.getIFDRationalValue method rightly returns null. But then the
IFD.get[XY]Resolution methods coerce the result to 0 and return it.
And then the downstream TIFFFormat code in complains about
non-positive values.

The fix was easy: return null instead of 0 when nothing is there.