Copy ImageMetadata upon translation
When translating from source to destination metadata, the ImageMetadata
of the source was not being preserved. But it certainly should be.

Of course, each format-specific translator is free to later tweak the
destination's ImageMetadata. But total loss of the source ImageMetadata
seems like a really bad plan, from my (admittedly naive) viewpoint.