TIFFFormat: improve multi-plane TIFF reading
With the merge of the 'tiff-fixes' branch, SCIFIO now records
the multidimensional structure when writing, and respects it
when reading. Unfortunately, this change resulted in a serious
regression: when such metadata is missing, it now only reads
the first image plane and ignores the rest.

This commit fixes that problem, by create an axis of type unknown
for the third dimension, with length equal to the number of IFDs.

Signed-off-by: Curtis Rueden <ctrueden@wisc.edu>