FormatService: fix race condition in init
The DefaultFormatService spawned a new thread to preload its data
structures. But this thread started immediately, and was so eager it
could get to the Format plugin injection step before the ServiceLoader
had gotten to the point of adding the FormatService to the index.
Hence, injection of Format plugins would fail because the FormatService
in the context was still unavailable/null.

The fix is to wait until the FormatService is available in the context
before proceeding with Format plugin creation and injection.