Update ImgOpener and IO API for new ImgFactory
The ImgFactory now knows its type, so we don't need to pass one along
side it. This simplifies the API of anything accepting an ImgFactory.
We keep the old methods, deprecated, for backwards compatibility.

This commit also cleans up generics-related raw type weirdness,
and inconsistencies between types needing to be RealType alone
versus RealType & NativeType. In actuality, the NativeType
requirement ripples all the way down. This change will probably
break some downstream code, but it needs to be done, and I do not
see a super easy way to preserve 100% source compatibility here.