Use Interval bounds instead of long array pairs
Using ImgLib2 Interval objects is more object-oriented, more type-safe,
and most crucially, does not confuse (min, max) with (offset, length).
We can query the Interval object for min, max and dimension (i.e.
length) at will, making the code simpler and easier to understand.

Before this change, some areas of the codebase (e.g. openPlane) used
(min, max) pairs, while others (e.g. private read methods) operated on
(offset, length) pairs. And in code that crossed this threshold, the
conversion was not always being done.

Without scrutinizing the revision history in detail, I am guessing that
the openPlane method did also operate on (offset, length) pairs at one
point, since the variables are still named that way in some places
(e.g. the ReaderFilter implementations). This was quite misleading
with (min, max)-related logic being performed on variables named
planeOffsets and planeLengths. And vice versa: many areas of code
treated variables named max or planarMax as dimensional lengths,
without first subtracting the bound's minimum. In the common case of
min=0 it works, but it fails for subregions not anchored at the origin.

And there was another oddity: planeMin was inclusive, while planeMax
was exclusive. This is not uncommon with discrete bounds, since it has
the nice property that length=max-min, rather than length=max-min+1.
However, ImgLib2's Interval uses inclusive min and max for consistency,
so this update does it that way as well.

Unfortunately, this change destroys backwards compatibility.
But given the magnitude of the inconsistency, it is worth it.