Remove openThumbPlane API
Nothing in core projects downstream requires it.
And it is an extra complicated wrinkle to maintain.

It should not be SCIFIO's job to manufacture thumbnails.
For formats that already have thumbnails, this can be
considered a second resolution, and can be supported
in the future via a true multi-resolution API.