Add FormatTools.rasterToPosition with expect. dims
FormatTools.rasterToPosition() computes a position array whose length is
based on the effective (i.e. the non-zero, non-trailing) axes as
extracted from image metadata. This commits adds another signature to
allow passing in a list of expected axes. If the image metadata doesn't
contain an axis that is expected, the entry in the array will be 0.