TiffSaver: Ensure header is written
The previously existing check in TIFFFormat was deciding to write the
header based on the length of the DataHandle. But the header should be
written..
	* in case the file does not exist yet
	* in case the file exists but is empty
	* in case the file is not empty (because the existing header
might be invalid / wrong)
Therefore this check makes no sense.
I added a check for the planeIndex because I assume it only makes sense
to write the header once. planeIndex == 0 should cover that.

The changes in TiffSaver:writeHeader is neccessary because out.seek(0)
is internally calling the reader() of the DataHandle - which only exists
if the file exists (see #405)