TIFFFormat: add ZLIB compression option
TiffCompression.DEFLATE is supported already, but wasn't exposed in TIFFFormat yet.

This commit adds a test similar to the others in TiffFormatTest.
Test coverage of the ZlibCodec class is confirmed by JaCoCo. In a future commit, we should add a check for the actual IFD tag in the written files, to ensure we actually use the specified compression.