Fix FileToDatasetConverter picking up non-existing files
It turns out that io.canOpen actually returns true for a file like 'blobs.gif' even if it doesn't exist on the given path. FileToDatasetConverter should really only be accepting requests where the file exists (or the name ends with '.fake' for the special case of the fake file format).

The tests are adjusted to be testing for unsupported formats as well as non-existing but theoretically supported files.