Make the RefCleaner a daemon thread
The ThreadService.run method launches non-daemon threads from the thread
pool. We need this RefCleaner to be a daemon thread (see #486) so that
it doesn't prevent JVM shutdown, and we also need it to start working
immediately and not be beholden to the ThreadService's associated
ExecutorService. Let's just spin our own Thread ourselves, it's fine.