Switch to jai-imageio-jpeg2000
I am hopeful that scifio-jai-imageio is obsolete now, due to
https://github.com/jai-imageio, which has existed for more than five
years now. This org is the closest thing we have to an "official"
revived jai-imageio project, so we should use it if possible rather than
rolling our own.

The slightly different values in the test are unfortunate, but it is
within some reasonable epsilon (http://tom7.org/papers/epsilon.pdf),
so I'm just going ahead with the switch. We can always revert back if
it causes disruptions downstream.